+++ src/main/kotlin/org/jetbrains/intellij/model/ObjectFactory.kt
@@ -8,5 +8,5 @@ class ObjectFactory {
 
     fun createPluginsCache() = PluginsCache()
 
-    fun createPluginVerifierRepository() = PluginVerifierRepository()
+    fun createPluginVerifierRepository() = SpacePackagesMavenMetadata()
 }