+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Use JDOM for altering `updates.xml` in `PrepareSandboxTask` to keep existing content
 - Fixed incorrect output path of JarSearchableOptionsTask causing also duplicate entry exception [#678](../../issues/678)
 - Fixed incorrect plugin download URL for custom repositories [#688](../../issues/688)
+- Make `DownloadRobotServerPluginTask` pointing to the latest Robot Server Plugin available
 
 ## 1.0.0